fix: square root test

+++ src/CalculatorTests.py
@@ -37,8 +37,8 @@ class MyTestCase(unittest.TestCase):
     def test_division(self):
         test_data = CsvReader('/src/Unit Test Division.csv').data
         for row in test_data:
-            self.assertEqual(self.calculator.division(row['Value 1'], row['Value 2']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.division(row['Value 1'], row['Value 2']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
 
     def test_square(self):
         test_data = CsvReader('/src/Unit Test Square.csv').data
@@ -49,8 +49,9 @@ class MyTestCase(unittest.TestCase):
     def test_root(self):
         test_data = CsvReader('/src/Unit Test Square Root.csv').data
         for row in test_data:
-            self.assertEqual(self.calculator.sqroot(row['Value 1'], row['Value 2']), int(row['Result']))
-            self.assertEqual(self.calculator.result, int(row['Result']))
+            self.assertEqual(self.calculator.sqroot(row['Value 1'], row['Value 2']), float(row['Result']))
+            self.assertEqual(self.calculator.result, float(row['Result']))
+
 
 if __name__ == '__main__':
     unittest.main()
